wallapop-autobuter/requirements.txt
2captcha-python==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
cloudscraper==1.2.71
dnspython==2.7.0
docker==7.1.0
fake-useragent==2.2.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
kafka-python>=2.0.0
outcome==1.3.0.post0
pyaes==1.6.1
pyee==13.0.0
PyJWT==2.10.1
pymongo==4.14.1
pyparsing==3.2.3
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.4
requests-toolbelt==1.0.0
selenium==4.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
